+++ dotCMS/src/integration-test/java/com/dotcms/MainSuite.java
@@ -166,8 +166,7 @@ import com.dotmarketing.startup.runonce.Task220402UpdateDateTimezonesTest;
 import com.dotmarketing.startup.runonce.Task220413IncreasePublishedPushedAssetIdColTest;
 import com.dotmarketing.startup.runonce.Task220512UpdateNoHTMLRegexValueTest;
 import com.dotmarketing.startup.runonce.Task220606UpdatePushNowActionletNameTest;
-import com.dotmarketing.startup.runonce.Task222208CreateVariantTableTest;
-import com.dotmarketing.startup.runonce.Task222408CreateDefaultVariantTest;
+import com.dotmarketing.startup.runonce.Task220824CreateDefaultVariantTest;
 import com.dotmarketing.util.ConfigTest;
 import com.dotmarketing.util.HashBuilderTest;
 import com.dotmarketing.util.MaintenanceUtilTest;
@@ -567,7 +566,7 @@ import org.junit.runners.Suite.SuiteClasses;
         VariantFactoryTest.class,
         VariantAPITest.class,
         PaginatedContentletsIntegrationTest.class,
-        Task222408CreateDefaultVariantTest.class
+        Task220824CreateDefaultVariantTest.class
 })
 public class MainSuite {
